+++ test/unittests/mongoBackend/mongoUpdateContextSubscription_test.cpp
@@ -234,7 +234,7 @@ static void prepareDatabase(void) {
 
     BSONObj sub1 = BSON("_id" << OID("51307b66f481db11bf860001") <<
                         "expiration" << 10000000 <<
-                        "lastNotification" << (long long) 15000000 <<
+                        "lastNotification" << 15000000 <<
                         "throttling" << 60 <<
                         "reference" << "http://notify1.me" <<
                         "format" << "XML" <<
@@ -252,7 +252,7 @@ static void prepareDatabase(void) {
 
     BSONObj sub2 = BSON("_id" << OID("51307b66f481db11bf860002") <<
                         "expiration" << 20000000 <<
-                        "lastNotification" << (long long) 25000000 <<
+                        "lastNotification" << 25000000 <<
                         "reference" << "http://notify2.me" <<
                         "format" << "XML" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false")) <<
@@ -270,7 +270,7 @@ static void prepareDatabase(void) {
     /* No type variant of sub2 */
     BSONObj sub2nt = BSON("_id" << OID("51307b66f481db11bf860022") <<
                         "expiration" << 20000000 <<
-                        "lastNotification" << (long long) 25000000 <<
+                        "lastNotification" << 25000000 <<
                         "reference" << "http://notify2.me" <<
                         "format" << "XML" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "isPattern" << "false")) <<
@@ -287,7 +287,7 @@ static void prepareDatabase(void) {
 
     BSONObj sub3 = BSON("_id" << OID("51307b66f481db11bf860003") <<
                         "expiration" << 30000000 <<
-                        "lastNotification" << (long long) 35000000 <<
+                        "lastNotification" << 35000000 <<
                         "reference" << "http://notify3.me" <<
                         "format" << "XML" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false") <<
@@ -305,7 +305,7 @@ static void prepareDatabase(void) {
 
     BSONObj sub4 = BSON("_id" << OID("51307b66f481db11bf860004") <<
                         "expiration" << 40000000 <<
-                        "lastNotification" << (long long) 45000000 <<
+                        "lastNotification" << 45000000 <<
                         "reference" << "http://notify4.me" <<
                         "format" << "XML" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false") <<
@@ -322,7 +322,7 @@ static void prepareDatabase(void) {
                         );
     BSONObj sub5 = BSON("_id" << OID("51307b66f481db11bf860005") <<
                         "expiration" << 50000000 <<
-                        "lastNotification" << (long long) 55000000 <<
+                        "lastNotification" << 55000000 <<
                         "throttling" << 10 <<
                         "reference" << "http://notify5.me" <<
                         "format" << "XML" <<
@@ -432,7 +432,7 @@ static void prepareDatabasePatternTrue(void) {
 
     BSONObj sub1 = BSON("_id" << OID("51307b66f481db11bf860001") <<
                         "expiration" << 10000000 <<
-                        "lastNotification" << (long long) 15000000 <<
+                        "lastNotification" << 15000000 <<
                         "throttling" << 60 <<
                         "reference" << "http://notify1.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false")) <<
@@ -449,7 +449,7 @@ static void prepareDatabasePatternTrue(void) {
 
     BSONObj sub2 = BSON("_id" << OID("51307b66f481db11bf860002") <<
                         "expiration" << 20000000 <<
-                        "lastNotification" << (long long) 25000000 <<
+                        "lastNotification" << 25000000 <<
                         "reference" << "http://notify2.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E[1-2]" << "type" << "T" << "isPattern" << "true")) <<
                         "attrs" << BSON_ARRAY("A1" << "A2") <<
@@ -466,7 +466,7 @@ static void prepareDatabasePatternTrue(void) {
     /* No type variant of sub2 */
     BSONObj sub2nt = BSON("_id" << OID("51307b66f481db11bf860022") <<
                         "expiration" << 20000000 <<
-                        "lastNotification" << (long long) 25000000 <<
+                        "lastNotification" << 25000000 <<
                         "reference" << "http://notify2.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E[1-2]" << "isPattern" << "true")) <<
                         "attrs" << BSON_ARRAY("A1" << "A2") <<
@@ -482,7 +482,7 @@ static void prepareDatabasePatternTrue(void) {
 
     BSONObj sub3 = BSON("_id" << OID("51307b66f481db11bf860003") <<
                         "expiration" << 30000000 <<
-                        "lastNotification" << (long long) 35000000 <<
+                        "lastNotification" << 35000000 <<
                         "reference" << "http://notify3.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false") <<
                                                  BSON("id" << "E2" << "type" << "T2" << "isPattern" << "false")) <<
@@ -499,7 +499,7 @@ static void prepareDatabasePatternTrue(void) {
 
     BSONObj sub4 = BSON("_id" << OID("51307b66f481db11bf860004") <<
                         "expiration" << 40000000 <<
-                        "lastNotification" << (long long) 45000000 <<
+                        "lastNotification" << 45000000 <<
                         "reference" << "http://notify4.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false") <<
                                                  BSON("id" << "E2" << "type" << "T2" << "isPattern" << "false")) <<
@@ -515,7 +515,7 @@ static void prepareDatabasePatternTrue(void) {
                         );
     BSONObj sub5 = BSON("_id" << OID("51307b66f481db11bf860005") <<
                         "expiration" << 50000000 <<
-                        "lastNotification" << (long long) 55000000 <<
+                        "lastNotification" << 55000000 <<
                         "throttling" << 10 <<
                         "reference" << "http://notify5.me" <<
                         "entities" << BSON_ARRAY(BSON("id" << "E1" << "type" << "T1" << "isPattern" << "false")) <<
